requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pycodestyle==2.13.0
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
random_word==1.0.13
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
